Made change for PR
Renamed file names and removed code
